Double definition error from Utils solved
